+++ src/types/associations/shared.ts
@@ -92,7 +92,7 @@ type NonKyselySupportedSupplementalWhereClauseValues<
     ? ColumnType extends 'bigint'
       ? TypesAllowedForBigintAgainstTheDb | PermanentOpsValTypes
       : ModelPropertyType extends DateTime | CalendarDate
-        ? DateTime | CalendarDate
+        ? DateTime | CalendarDate | null
         : ModelPropertyType extends number | string
           ? ModelPropertyType | PermanentOpsValTypes
           : never